+++ cli_tool_audit/__main__.py
@@ -4,26 +4,33 @@ Argument parsing code.
 import argparse
 import logging
 import logging.config
-import os
 import sys
+from collections.abc import Sequence
 from dataclasses import fields
-from typing import Any
+from typing import Any, Optional
 
+import cli_tool_audit.config_manager as config_manager
+import cli_tool_audit.freeze as freeze
+import cli_tool_audit.interactive as interactive
+import cli_tool_audit.logging_config as logging_config
+import cli_tool_audit.view_npm_stress_test as demo_npm
+import cli_tool_audit.view_pipx_stress_test as demo_pipx
+import cli_tool_audit.view_venv_stress_test as demo_venv
+import cli_tool_audit.views as views
 from cli_tool_audit._version import __version__
-from cli_tool_audit.config_manager import CliToolConfig, ConfigManager
-from cli_tool_audit.freeze import freeze_to_screen
-from cli_tool_audit.view_npm_stress_test import report_for_npm_tools
-from cli_tool_audit.view_pipx_stress_test import report_for_pipx_tools
-from cli_tool_audit.view_venv_stress_test import report_for_venv_tools
-from cli_tool_audit.views import report_from_pyproject_toml
 
 logger = logging.getLogger(__name__)
 
 
-def handle_read(args):
-    config_manager = ConfigManager(args.config)
-    config_manager.read_config()
-    for tool, config in config_manager.tools.items():
+def handle_read(args: argparse.Namespace) -> None:
+    """
+    Read and list all tool configurations
+    Args:
+        args: The args from the command line.
+    """
+    manager = config_manager.ConfigManager(args.config)
+    manager.read_config()
+    for tool, config in manager.tools.items():
         print(f"{tool}")
         for key, value in vars(config).items():
             if key == "only_check_existence" and value is False:
@@ -32,53 +39,137 @@ def handle_read(args):
                 print(f"  {key}: {value}")
 
 
-def handle_create(args):
+def handle_create(args: argparse.Namespace) -> None:
+    """
+    Create a new tool configuration.
+    Args:
+        args: The args from the command line.
+    """
     kwargs = reduce_args_tool_cli_tool_config_args(args)
 
-    config_manager = ConfigManager(args.config)
-    config_manager.create_tool_config(args.tool, kwargs)
+    manager = config_manager.ConfigManager(args.config)
+    manager.create_tool_config(args.tool, kwargs)
     print(f"Tool {args.tool} created.")
 
 
-def reduce_args_tool_cli_tool_config_args(args):
+def reduce_args_tool_cli_tool_config_args(args: argparse.Namespace) -> dict[str, Any]:
+    """
+    Reduce the args to only those that are in CliToolConfig.
+    Args:
+        args: The args from the command line.
+
+    Returns:
+        dict: The reduced args.
+    """
     kwargs = {}
     args_dict = vars(args)
-    cli_tool_fields = {f.name for f in fields(CliToolConfig)}
+    cli_tool_fields = {f.name for f in fields(config_manager.CliToolConfig)}
     for key, value in args_dict.items():
         if key in cli_tool_fields:
             kwargs[key] = value
     return kwargs
 
 
-def handle_update(args):
+def handle_update(args: argparse.Namespace) -> None:
     kwargs = reduce_args_tool_cli_tool_config_args(args)
-    config_manager = ConfigManager(args.config)
-    config_manager.update_tool_config(args.tool, {k: v for k, v in kwargs.items() if k != "tool"})
+    manager = config_manager.ConfigManager(args.config)
+    manager.update_tool_config(args.tool, {k: v for k, v in kwargs.items() if k != "tool"})
     print(f"Tool {args.tool} updated.")
 
 
-def handle_delete(args):
-    config_manager = ConfigManager(args.config)
-    config_manager.delete_tool_config(args.tool)
+def handle_delete(args: argparse.Namespace) -> None:
+    """
+    Delete a tool configuration.
+    Args:
+        args: The args from the command line.
+    """
+    manager = config_manager.ConfigManager(args.config)
+    manager.delete_tool_config(args.tool)
     print(f"Tool {args.tool} deleted.")
 
 
-def main() -> None:
-    """Parse arguments and run the CLI tool."""
+def handle_interactive(args: argparse.Namespace) -> None:
+    """
+    Interactively edit configuration from terminal.
+    Args:
+        args: The args from the command line.
+    """
+    manager = config_manager.ConfigManager(args.config)
+    interactive.interactive_config_manager(manager)
+
+
+def add_update_args(parser: argparse.ArgumentParser) -> None:
+    """
+    Add arguments to the add or update parser.
+    Args:
+        parser: The add or update parser.
+    """
+    parser.add_argument("--version", help="Version of the tool")
+    parser.add_argument("--version-switch", "--version_switch", help="Version switch for the tool")
+    parser.add_argument("--version-snapshot", "--version_snapshot", help="Entire capture of version is version string.")
+    parser.add_argument("--if-os", "--if_os", help="Check only on this os.")
+    parser.add_argument(
+        "--only-check-existence",
+        "--only_check_existence",
+        action="store_true",
+        help="Check only the existence of the tool",
+    )
+    # TODO: Add more arguments
+
+
+def main(argv: Optional[Sequence[str]] = None) -> int:
+    """Parse arguments and run the CLI tool.
+    Args:
+        argv: The arguments to parse.
+
+    Returns:
+        int: The exit code.
+    """
     # Create the parser
-    parser = argparse.ArgumentParser(description="Audit version numbers of CLI tools.")
-    parser.add_argument(
-        "--version", action="version", version=f"%(prog)s {__version__}", help="Show program's version number and exit."
+    parser = argparse.ArgumentParser(
+        prog="cli-tool-audit", allow_abbrev=False, description="Audit version numbers of CLI tools."
     )
     parser.add_argument(
-        "--config", default="pyproject.toml", type=str, help="Path to the configuration file in TOML format."
+        "-V",
+        "--version",
+        action="version",
+        version=f"%(prog)s {__version__}",
+        help="Show program's version number and exit.",
     )
-    parser.add_argument("--never_fail", action="store_true", help="Never return a non-zero exit code")
+    parser.add_argument(
+        "-c",
+        "--config",
+        default="pyproject.toml",
+        type=str,
+        help="Path to the configuration file in TOML format. (default is %(default)s)",
+    )
+    parser.add_argument(
+        "-ff",
+        "--file-format",
+        "--file_format",
+        default="table",
+        type=str,
+        choices=("json", "json-compact", "xml", "table", "csv"),
+        help="Output results in the specified format. (default is %(default)s)",
+    )
+
+    parser.add_argument(
+        "-nf", "--never-fail", "--never_fail", action="store_true", help="Never return a non-zero exit code"
+    )
     parser.add_argument("--verbose", action="store_true", help="verbose output")
 
-    parser.add_argument("--demo", type=str, help="Demo for values of npm, pipx or venv")
+    parser.add_argument(
+        "--demo",
+        type=str,
+        choices=("pipx", "venv", "npm"),
+        help="Demo for values of npm, pipx or venv",
+    )
 
-    subparsers = parser.add_subparsers(help="commands")
+    subparsers = parser.add_subparsers(help="Edit configuration from terminal.")
+
+    # Read command
+    interactive_parser = subparsers.add_parser("interactive", help="Interactively edit configuration")
+    interactive_parser.set_defaults(func=handle_interactive)
 
     # Read command
     read_parser = subparsers.add_parser("read", help="Read and list all tool configurations")
@@ -87,30 +178,13 @@ def main() -> None:
     # Create command
     create_parser = subparsers.add_parser("create", help="Create a new tool configuration")
     create_parser.add_argument("tool", help="Name of the tool")
-    create_parser.add_argument("--version", help="Version of the tool")
-    create_parser.add_argument("--version_switch", nargs="?", help="Version switch for the tool")
-    create_parser.add_argument(
-        "--only_check_existence", action="store_true", help="Check only the existence of the tool"
-    )
-    # ... add other arguments as needed
+    add_update_args(create_parser)
     create_parser.set_defaults(func=handle_create)
 
     # Update command
     update_parser = subparsers.add_parser("update", help="Update an existing tool configuration")
     update_parser.add_argument("tool", help="Name of the tool")
-    update_parser.add_argument("--version", help="Version of the tool")
-    update_parser.add_argument("--version_switch", help="Version switch for the tool")
-    update_parser.add_argument("--version_snapshot", help="Entire capture of version is version string.")
-    update_parser.add_argument("--if_os", help="Check only on this os.")
-    update_parser.add_argument(
-        "--only_check_existence", action="store_true", help="Check only the existence of the tool"
-    )
-
-    # TODO:
-    # schema: Optional[str] = None
-    # if_os: Optional[str] = None
-    # version_stamp: Optional[str] = None
-    # source: Optional[str] = None
+    add_update_args(update_parser)
 
     update_parser.set_defaults(func=handle_update)
 
@@ -123,38 +197,17 @@ def main() -> None:
     freeze_parser = subparsers.add_parser("freeze", help="Freeze the versions of specified tools")
     freeze_parser.add_argument("tools", nargs="+", help="List of tool names to freeze")
 
+    # Add 'audit' sub-command
+    _audit_parser = subparsers.add_parser("audit", help="Audit environment with current configuration (default)")
+    # audit_parser.add_argument("tools", nargs="+", help="...")
+
+    # --------------------------------------------------------------------------
+
     # Parse the arguments
-    args = parser.parse_args()
+    args = parser.parse_args(argv)
 
     if args.verbose:
-        config: dict[str, Any] = {
-            "version": 1,
-            "disable_existing_loggers": True,
-            "formatters": {
-                "standard": {"format": "[%(levelname)s] %(name)s: %(message)s"},
-                "colored": {
-                    "()": "colorlog.ColoredFormatter",
-                    "format": "%(log_color)s%(levelname)-8s%(reset)s %(green)s%(message)s",
-                },
-            },
-            "handlers": {
-                "default": {
-                    "level": "DEBUG",
-                    "formatter": "colored",
-                    "class": "logging.StreamHandler",
-                    "stream": "ext://sys.stdout",  # Default is stderr
-                },
-            },
-            "loggers": {
-                "cli_tool_audit": {
-                    "handlers": ["default"],
-                    "level": "DEBUG",
-                    "propagate": False,
-                }
-            },
-        }
-        if os.environ.get("NO_COLOR"):
-            config["handlers"]["default"]["formatter"] = "standard"
+        config = logging_config.generate_config(level="DEBUG")
         logging.config.dictConfig(config)
     else:
         # Essentially, quiet mode
@@ -164,34 +217,31 @@ def main() -> None:
 
     # Demos
     if args.demo and args.demo == "pipx":
-        report_for_pipx_tools()
-        sys.exit()
+        demo_pipx.report_for_pipx_tools()
+        return 0
     if args.demo and args.demo == "venv":
-        report_for_venv_tools()
-        sys.exit()
+        demo_venv.report_for_venv_tools()
+        return 0
     if args.demo and args.demo == "npm":
-        report_for_npm_tools()
-        sys.exit()
+        demo_npm.report_for_npm_tools()
+        return 0
 
-    args = parser.parse_args()
     if hasattr(args, "func"):
         args.func(args)
-        sys.exit()
-    # else:
-    #     parser.print_help()
+        return 0
 
     # Namespace doesn't have word "freeze" in it.
     if hasattr(args, "tools") and args.tools:
-        freeze_to_screen(args.tools)
-        sys.exit()
+        freeze.freeze_to_screen(args.tools)
+        return 0
 
     # Default behavior
-    # Handle the configuration file argument
     if args.config:
-        report_from_pyproject_toml(args.config, exit_code_on_failure=not args.never_fail)
-    else:
-        report_from_pyproject_toml(exit_code_on_failure=not args.never_fail)
+        return views.report_from_pyproject_toml(
+            args.config, exit_code_on_failure=not args.never_fail, file_format=args.file_format
+        )
+    return views.report_from_pyproject_toml(exit_code_on_failure=not args.never_fail, file_format=args.file_format)
 
 
 if __name__ == "__main__":
-    main()
+    sys.exit(main())