+++ cli_tool_audit/view_venv_stress_test.py
@@ -11,8 +11,8 @@ from concurrent.futures import ThreadPoolExecutor
 # pylint: disable=no-name-in-module
 from whichcraft import which
 
-from cli_tool_audit.config_manager import CliToolConfig
-from cli_tool_audit.views import check_tool_wrapper, pretty_print_results
+import cli_tool_audit.views as views
+from cli_tool_audit.models import CliToolConfig
 
 
 def get_executables_in_venv(venv_path: str) -> list[str]:
@@ -38,31 +38,37 @@ def get_executables_in_venv(venv_path: str) -> list[str]:
     return executables
 
 
-def report_for_venv_tools() -> None:
+def report_for_venv_tools(max_count: int = -1) -> None:
     """
     Report on the compatibility of the tools installed in the virtual environment.
+    Args:
+        max_count (int, optional): The maximum number of tools to report on. Defaults to -1.
     """
     python_path = which("python")
     venv_dir = pathlib.Path(python_path).parent.parent
     cli_tools = {}
+    count = 0
     for executable in get_executables_in_venv(str(venv_dir)):
         config = CliToolConfig()
         config.version_switch = "--version"
         config.version = ">=0.0.0"
         cli_tools[executable] = config
+        count += 1
+        if count >= max_count > 0:
+            break
     # Determine the number of available CPUs
     num_cpus = os.cpu_count()
 
     # Create a ThreadPoolExecutor with one thread per CPU
     with ThreadPoolExecutor(max_workers=num_cpus) as executor:
         # Submit tasks to the executor
-        futures = [executor.submit(check_tool_wrapper, (tool, config)) for tool, config in cli_tools.items()]
+        futures = [executor.submit(views.check_tool_wrapper, (tool, config)) for tool, config in cli_tools.items()]
         results = []
         # Process the results as they are completed
         for future in concurrent.futures.as_completed(futures):
             result = future.result()
             results.append(result)
-        pretty_print_results(results)
+        views.pretty_print_results(results)
 
 
 if __name__ == "__main__":