+++ CHANGES.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-01-16
+
+### Added
+- Interactive config at cli with `--interactive`
+- Short switches for cli args and aliases for - and _ as connectors
+- Export multiple file formats with `--file-format`
+- Added subcommand for `audit` with intention of removing default action.
+
+### Fixed
+- Spelling and docs. Make file now lints docs and runs spell check tools.
+
 ## [1.0.7] - 2024-01-16
 
 ### Fixed